Renamed int to double

+++ PoGo.NecroBot.Logic/Settings.cs
@@ -650,7 +650,7 @@ namespace PoGo.NecroBot.Logic
         public bool UseTransferIvForSnipe => _settings.UseTransferIvForSnipe;
         public bool SnipeIgnoreUnknownIv => _settings.SnipeIgnoreUnknownIv;
         public int MinDelayBetweenSnipes => _settings.MinDelayBetweenSnipes;
-        public int SnipingScanOffset => _settings.SnipingScanOffset;
+        public double SnipingScanOffset => _settings.SnipingScanOffset;
         public int TotalAmountOfPokebalsToKeep => _settings.TotalAmountOfPokebalsToKeep;
         public int TotalAmountOfPotionsToKeep => _settings.TotalAmountOfPotionsToKeep;
         public int TotalAmountOfRevivesToKeep => _settings.TotalAmountOfRevivesToKeep;
